Rough but mostly finished version of mmgrid.py, with tests
This version is mostly feature complete, but the only tagging
mechanism available through the command-line script is a
visualization-style tagging similar to reTagMesh, and some
other minor features are missing.

The tests currently don't assume you have pydagmc available
on your PYTHONPATH and will be skipped if it's not found.