Code to write an alara geom file based on mmgrid output
This is essentially a replacement for moab2alara
compatible with mmgrid.py.  moab2alara.py is preserved
for now as a means of getting information from files
produced by mmGridGen.

This script still has a few loose ends and rough edges,
but the basic functionality is here.  It can convert a
single mmgrid output file passed to it on the command line.